LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssTransformComponent-2d-flattening-expected.txt


PASS CSSTranslate.toMatrix() flattens when told it is 2d
PASS CSSRotate.toMatrix() flattens when told it is 2d
PASS CSSScale.toMatrix() flattens when told it is 2d
PASS CSSMatrixComponent.toMatrix() flattens when told it is 2d
